+++ src/tracksdata/graph/_rustworkx_graph.py
@@ -10,8 +10,8 @@ import rustworkx as rx
 from tracksdata.attrs import AttrComparison, split_attr_comps
 from tracksdata.constants import DEFAULT_ATTR_KEYS
 from tracksdata.functional._rx import _assign_track_ids
-from tracksdata.graph._base_filter import BaseFilter, cache_method
 from tracksdata.graph._base_graph import BaseGraph
+from tracksdata.graph.filters._base_filter import BaseFilter, cache_method
 from tracksdata.utils._dataframe import unpack_array_attrs
 from tracksdata.utils._logging import LOG
 
@@ -1557,7 +1557,7 @@ class IndexedRXGraph(RustWorkXGraph):
         include_targets: bool = False,
         include_sources: bool = False,
     ) -> RXFilter:
-        from tracksdata.graph._indexed_filter import IndexRXFilter
+        from tracksdata.graph.filters._indexed_filter import IndexRXFilter
 
         return IndexRXFilter(
             *attr_filters,
