+++ src/tracksdata/graph/filters/_base_filter.py
